requirements.txt
flask
mysql-connector-python
pydantic
python-dotenv